pyproject.toml
[tool.poetry]
name = "logintest"
version = "0.1.0"
description = ""
authors = ["Mike <Walker>"]

[tool.poetry.dependencies]
python = "^3.6.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
